+++ src/KCC.Web/Features/Components/Breadcrumbs/Breadcrumbs.cshtml
@@ -1,0 +1,21 @@
+@using KCC.Web.Features.Components.Breadcrumbs
+@model BreadcrumbsViewModel
+
+<div data-ktc-search-exclude>
+  @foreach (var link in Model.Links)
+  {
+    if (link != Model.Links.First())
+    {
+      <span class="px-2">/</span>
+    }
+
+    @if (link.Url != string.Empty)
+    {
+      <a href="@link.Url" class="underline">@link.LinkText</a>
+    }
+    else
+    {
+      <span class="text-gray-500">@link.LinkText</span>
+    }
+  }
+</div>
