+++ src/KCC.Web/Features/Pages/ViewStart.cshtml
@@ -1,0 +1,3 @@
+@{
+  Layout = "~/Features/Pages/Shared/Layout.cshtml";
+}
