--- src/KCC.Web/Features/Pages/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-@{
-    Layout = "~/Features/Pages/Shared/_Layout.cshtml";
-}
