+++ src/KCC.Web/Features/Widgets/Hero/HeroWidget.cshtml
